% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{abcox_fit}
\alias{abcox_fit}
\title{Do the Approximate Bayesian Inference for the specified Cox Proportional Hazard Model.}
\usage{
abcox_fit(
  abcox_model,
  control = abcox_default_control(),
  PARALLEL_EXECUTION = T,
  method = ""
)
}
\arguments{
\item{abcox_model}{A result from abcoxph_setup}

\item{control}{A list from abcox_default_control}

\item{PARALLEL_EXECUTION}{Whether you want the computations to be done in parallel.}
}
\value{
A list that contains marginal_latent, marginal_hyper and the original model_data from your model.
}
\description{
Using the method described in the paper, to perform the approximate Bayesian inference for the specified CoxPH model,
under you specified controls.
}
