% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{abcoxph_AGHQ_setup}
\alias{abcoxph_AGHQ_setup}
\title{Set up the Cox Proportional Hazard Model that you want to infer with AGHQ method.}
\usage{
abcoxph_AGHQ_setup(
  formula,
  cens,
  data,
  prior_control,
  RW2BINS = NULL,
  method = "BFGS",
  inner_method = "trust",
  correction = "FALSE"
)
}
\arguments{
\item{formula}{A formula specifies your observed times as response variable, regressors type are smoothing, id, or linear fixed effect.}

\item{cens}{A STRING specifies the name of your censoring indicator in the dataframe.}

\item{data}{Target dataframe.}

\item{prior_control}{A list specifies your prior distributions, this must be generated from prior_setup_AGHQ.}

\item{RW2BINS}{A number specifies the number of bins you want to use for smoothing.}

\item{method}{The outer method used in AGHQ for the optimization, suggested to use "BFGS" at all the time.}

\item{inner_method}{The inner method used in AGHQ for the optimization, suggested to use "trust", if the dataset is large, suggested to use "SR1"}

\item{correction}{Whether to use the Breslow's method for the correction of ties, suggested when the dataset is small and number of ties large.}
}
\value{
A list that contains all the information needed for your model to be fitted using abcoxph_fit.
}
\description{
Setting up a target CoxPH Model and its AGHQ specification, by specifying the formula. The response in the variable should be your
observed times, and the regressors's type are well identified with s denote smoothing and id denote between-subject
index, other terms are assumed to be linear fixed effect.
}
