% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-likelihood.R
\name{Get_Observed_time}
\alias{Get_Observed_time}
\title{Get the observed times from a data set.}
\usage{
Get_Observed_time(model_data)
}
\arguments{
\item{model_data}{A list of class "abcox_modeldata" as returned by model_setup().}
}
\value{
A vector
}
\description{
Take a latent field vector and return the (transformed) linear predictors that we want.
}
