% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-latent-variables.R
\name{logprior_W}
\alias{logprior_W}
\alias{log_posterior_W}
\alias{grad_log_posterior_W}
\alias{hessian_log_posterior_W}
\title{log-prior of W}
\usage{
logprior_W(W, theta, model_data, Q = NULL)

log_posterior_W(W, theta, model_data, Q = NULL)

grad_log_posterior_W(W, theta, model_data, Q = NULL)

hessian_log_posterior_W(W, theta = NULL, Q = NULL, model_data)
}
\arguments{
\item{W}{Value of W = (delta,gamma,beta) to calculate the prior/posterior at. The order is as
it appears in the appendix of the paper- (delta_1_1,...,delta_n_Jn,gamma_1,...,gamma_M,beta_1,...,beta_p).
The order of gamma and beta is the same as they are listed in model_data$model_elements.}

\item{theta}{Value of the hyperparameter vector theta. The Q matrix depends on this. Can leave as NULL
if you're passing in the Q matrix}

\item{model_data}{Model's data created by abcoxph_setup}

\item{Q}{The Q-matrix as returned by Q_matrix(model). If not provided, will be calculated.}
}
\description{
These functions calculate priors and posteriors of the latent Gaussian variables.
}
