% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{plot_sigma}
\alias{plot_sigma}
\title{Plot the posterior for sigma}
\usage{
plot_sigma(abcox_fit)
}
\arguments{
\item{abcox_fit}{A result from abcox_fit}
}
\value{
A Plot for sigma's posterior.
}
\description{
Plot the posterior for sigma i.e.the standard deviation, which is attained using laplace approximation with
a transformation.
}
