% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{plot_smoothing}
\alias{plot_smoothing}
\title{Plot the inferred smoothing function.}
\usage{
plot_smoothing(abcox_fit)
}
\arguments{
\item{abcox_fit}{A result from abcox_fit}
}
\value{
A Plot for the smoothing result.
}
\description{
Plot the posterior mean of the smoothing function, and its corresponding credible interval.
}
