% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{plot_theta}
\alias{plot_theta}
\title{Plot the posterior for theta.}
\usage{
plot_theta(abcox_fit)
}
\arguments{
\item{abcox_fit}{A result from abcox_fit}
}
\value{
A Plot for theta's posterior.
}
\description{
Plot the posterior for theta i.e.log-precision, which is attained using laplace approximation.
}
