% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-likelihood.R
\name{prep_data_for_log_lik}
\alias{prep_data_for_log_lik}
\title{Prepare the data for computing the log-likelihood}
\usage{
prep_data_for_log_lik(W, model_data)
}
\arguments{
\item{W}{Parameter vector. First n-1 elements are (transformed) linear predictors, then Gamma and beta.}

\item{model_data}{A list of class "abcox_modeldata" as returned by model_setup().}
}
\value{
A vector of transformed linear predictors.
}
\description{
Take a latent field vector and return the (transformed) linear predictors that we want.
}
