% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{prior_setup}
\alias{prior_setup}
\title{Set up the prior distributions that you want to use.}
\usage{
prior_setup(
  theta_dim = 1,
  theta_number = 50,
  theta_grid = c(-2, 18),
  theta_prior_a = 0.5,
  theta_prior_u = 2,
  beta_logprec = log(0.001)
)
}
\arguments{
\item{theta_dim}{How many hyperparameters do you have?}

\item{theta_number}{For each parameter, how many grids you want to use for the integration?}

\item{theta_grid}{Your specific grid for theta to be used.}

\item{theta_prior_a}{A values for the pc.prior, the probability}

\item{theta_prior_u}{A values for the pc.prior, the lower bound}

\item{beta_logprec}{A values for the log precision of you fixed effect parameter.}
}
\value{
A list that contains all the information needed for your prior to be fitted using abcoxph_setup.
}
\description{
Setting up your prior distributions to be used.
}
