% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{useAGHQ}
\alias{useAGHQ}
\title{Do the AGHQ to find the theta_grid}
\usage{
useAGHQ(
  model_data,
  k = 3,
  startingvals = 0,
  method = "BFGS",
  inner_method = "trust"
)
}
\arguments{
\item{model_data}{the model_data created by abcoxph_setup or abcoxph_AGHQ_setup}

\item{k}{number of points in the grid}

\item{startingvals}{the starting value of theta in optimization}
}
\value{
An NIGrid object of theta_grid
}
\description{
Implement AGHQ to optimize for theta_grid
}
